Project changes - Added the real support for js and native platforms - Removed the Hex object to avoid unnecessary overhead - Renamed the `khex` module into `core` - Renamed the `khex-extensions` module into `extensions` - Changed the project group to `org.komputing.khex`
