+++ .gitignore
@@ -1,4 +1,3 @@
-/build
+build/
 .idea
-*.iml
 local.properties
