+++ gradle.properties
@@ -2,7 +2,7 @@
 kotlin.code.style=official
 
 # config
-version=1.0.0-RC1
+version=0.9
 kotlin.incremental.multiplatform=true
 
 # gradle
